+++ benchmarks/golden_configs/lm_wikitext2.py
@@ -80,19 +80,12 @@ class Pipe:
             "criterion": nn.CrossEntropyLoss(),
         }
 
-    def get_golden_real_stats(multiprocess=False):
-        if not multiprocess:
-            return {
-                "avg_wps": 703.778,
-                "std_dev_wps": 5.732,
-                "peak_mem_usage": [2320996352, 1396742144, 1396742144, 2340010496],
-            }
-        else:
-            return {
-                "avg_wps": 647.404,
-                "std_dev_wps": 14.51,
-                "peak_mem_usage": [3305007616, 2578692608, 3304524288, 2578692608],
-            }
+    def get_golden_real_stats():
+        return {
+            "avg_wps": 703.778,
+            "std_dev_wps": 5.732,
+            "peak_mem_usage": [2320996352, 1396742144, 1396742144, 2340010496],
+        }
 
     def get_golden_synthetic_stats():
         # TODO(anj-s): Add support for synthetic regression benchmarks