+++ docs/source/conf.py
@@ -29,7 +29,7 @@ copyright = "2020-2021, Facebook AI Research"
 author = "Facebook AI Research"
 
 # The full version, including alpha/beta/rc tags
-release = "0.3.2"
+release = "0.3.3"
 
 
 # -- General configuration ---------------------------------------------------