+++ fairscale/nn/pipe/__init__.py
@@ -20,7 +20,6 @@
 """A Pipe implementation in PyTorch."""
 from .async_pipe import AsyncPipe
 from .checkpoint import is_checkpointing, is_recomputing
-from .multiprocess_pipe import LazyModule, MultiProcessPipe
 from .pipe import Pipe
 from .rpc import PipeRPCWrapper